requirements.txt
fastapi
pydantic-settings
uvicorn
sqlalchemy
asyncpg
aiomqtt
pynacl
passlib[bcrypt]
python-jose[cryptography]